Clean out the runFlatten.ts
