+++ packages/discovery/src/utils/fsLayer.ts
@@ -1,5 +1,6 @@
 import { readdirSync } from 'fs'
-import { basename, dirname } from 'path'
+import { readdir } from 'fs/promises'
+import { basename, dirname, resolve } from 'path'
 
 // NOTE(radomski): On some file systems, mainly Apple's AFS and Microsoft's
 // NTFS the path names are not case sensitive. So while you can have a file
@@ -17,4 +18,16 @@ export function fileExistsCaseSensitive(path: string): boolean {
   }
 
   return true
+}
+
+export async function listFilesRecursively(path: string): Promise<string[]> {
+  const dirents = await readdir(path, { withFileTypes: true })
+  const files = await Promise.all(
+    dirents.map((dirent) => {
+      const res = resolve(path, dirent.name)
+      return dirent.isDirectory() ? listFilesRecursively(res) : res
+    }),
+  )
+
+  return files.flat()
 }
